+++ frontend/src/App/Layout/Header/OnlyDustLogo/OnlyDustLogo.stories.tsx
